ci speedup cleanup

+++ bin/travis-build.bash
@@ -5,43 +5,55 @@ echo "This is travis-build.bash..."
 
 echo "Installing the packages that CKAN requires..."
 sudo apt-get update -qq
-sudo apt-get install solr-jetty libcommons-fileupload-java
+sudo apt-get install -y solr-jetty libcommons-fileupload-java
+
+# !/bin/bash
+ver=$(python -c"import sys; print(sys.version_info.major)")
+if [ $ver -eq 2 ]; then
+    echo "python version 2"
+elif [ $ver -eq 3 ]; then
+    echo "python version 3"
+else
+    echo "Unknown python version: $ver"
+fi
 
 echo "Installing CKAN and its Python dependencies..."
 if [ ! -d ckan ]; then
-  git clone https://github.com/$CKAN_GIT_REPO
+
+  echo "Creating the PostgreSQL user and database..."
+  sudo -u postgres psql -c "CREATE USER ckan_default WITH PASSWORD 'pass';"
+  sudo -u postgres psql -c 'CREATE DATABASE ckan_test WITH OWNER ckan_default;'
+
+  if [ "${CKAN_BRANCH}dd" == 'dd' ]; then
+    #remote lookup tags, and get latest by version-sort
+    CKAN_TAG=$(git ls-remote --tags https://github.com/$CKAN_GIT_REPO | grep refs/tags/ckan-$CKANVERSION | awk '{print $2}'| sort --version-sort | tail -n 1 | sed  's|refs/tags/||' )
+    echo "CKAN tag version $CKANVERSION is: ${CKAN_TAG#ckan-}"
+    git clone --depth=50 --branch=$CKAN_TAG https://github.com/$CKAN_GIT_REPO ckan
+  else
+    echo "CKAN version: $CKAN_BRANCH"
+    git clone --depth=50 --branch=$CKAN_BRANCH https://github.com/$CKAN_GIT_REPO ckan
+  fi
 fi
+
 pushd ckan
 
-#are we building from branch or tag
-if [ "${CKAN_BRANCH}dd" == 'dd' ]
-then
-    CKAN_TAG=$(git tag | grep ^ckan-$CKANVERSION | sort --version-sort | tail -n 1)
-    git checkout $CKAN_TAG
-    echo "CKAN version: ${CKAN_TAG#ckan-}"
-else
-    echo "CKAN version: $CKAN_BRANCH"
-    git checkout $CKAN_BRANCH
+if [ $ver -eq 3 ]; then
+pip install -r requirement-setuptools.txt
 fi
 
-python setup.py develop
-if [ -f requirements-py2.txt ]
-then
+if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]; then
     pip install -r requirements-py2.txt
 else
     pip install -r requirements.txt
 fi
 pip install -r dev-requirements.txt
-popd
-
-echo "Creating the PostgreSQL user and database..."
-sudo -u postgres psql -c "CREATE USER ckan_default WITH PASSWORD 'pass';"
-sudo -u postgres psql -c 'CREATE DATABASE ckan_test WITH OWNER ckan_default;'
+python setup.py develop
 
 echo "Initialising the database..."
-pushd ckan
 paster db init -c test-core.ini
+
 popd
+
 
 echo "SOLR config..."
 # solr is multicore for tests on ckan master now, but it's easier to run tests
@@ -49,17 +61,32 @@ echo "SOLR config..."
 # see https://github.com/ckan/ckan/issues/2972
 sed -i -e 's/solr_url.*/solr_url = http:\/\/127.0.0.1:8983\/solr/' ckan/test-core.ini
 
-#echo "Installing dependency ckanext-report and its requirements..."
-#pip install -e git+https://github.com/datagovuk/ckanext-report.git#egg=ckanext-report
-
 echo "Installing ckanext-archiver and its requirements..."
-python setup.py develop
-pip install -r requirements.txt
+
+if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]; then
+  pip install -r requirements-py2.txt
+elif [ -f requirements.txt ]; then
+  pip install -r requirements.txt
+fi
 pip install -r dev-requirements.txt
+python setup.py develop
+
+echo "Installing dependency ckanext-report and its requirements..."
+if [ ! -d ckanext-report ]; then
+  git clone --depth=50 --branch=$REPORT_BRANCH https://github.com/$REPORT_GIT_REPO/ckanext-report ckanext-report
+fi
+pushd ckanext-report
+  if [ -f requirements-py2.txt ] && [ $ver -eq 2 ]; then
+    pip install -r requirements-py2.txt
+  elif [ -f requirements.txt ]; then
+    pip install -r requirements.txt
+  fi
+  pip install --no-deps -e .
+popd
 
 echo "Moving test-core.ini into a subdir..."
-mkdir subdir
-mv test-core.ini subdir
+mkdir -p subdir
+cp test-core.ini subdir
 
 echo "start solr"
 # Fix solr-jetty starting issues https://stackoverflow.com/a/56007895
@@ -75,7 +102,6 @@ sudo service jetty9 restart
 
 # Wait for jetty9 to start
 timeout 20 bash -c 'while [[ "$(curl -s -o /dev/null -I -w %{http_code} http://localhost:8983)" != "200" ]]; do sleep 2;done'
-
 
 
 echo "travis-build.bash is done."
